CHET-594: [frontend] separate routing from router and use redirect hook at root of router
