+++ frontend/src/main/dc-migration-assistant-fe/components/MigrationAssistant.tsx
@@ -15,8 +15,13 @@
  */
 
 import React, { FunctionComponent } from 'react';
-import { Routes } from './Routes';
+import { BrowserRouter as Router } from 'react-router-dom';
+import { Routes } from '../routing/Routes';
 
 export const MigrationAssistant: FunctionComponent = () => {
-    return <Routes />;
+    return (
+        <Router>
+            <Routes />
+        </Router>
+    );
 };
